55 입고처리 해결

+++ src/main/java/com/itwillbs/domain/MaterialOrderItemVO.java
@@ -1,6 +1,5 @@
 package com.itwillbs.domain;
 
-import java.math.BigDecimal;
 import lombok.Data;
 
 @Data
@@ -10,8 +9,8 @@ public class MaterialOrderItemVO {
     private String orderId;               // 발주 ID (연결된 발주서 ID)
     private String materialId;            // 자재 ID (발주한 자재 ID)
     private int orderQuantity;            // 수량 (요청한 수량)
-    private BigDecimal unitPrice;         // 단가 (단가)
-    private BigDecimal totalPrice;        // 총 금액 (수량 * 단가)
+    private int unitPrice;         		  // 단가 (단가)
+    private int totalPrice;        		   // 총 금액 (수량 * 단가)
     private String warehouseCode;         // 입고창고 (자재 납입 예정 창고)
     
     // 화면 표시용 추가 필드
